% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{available_genomes}
\alias{available_genomes}
\title{A data frame of genome info from UCSC}
\format{
An object of class \code{data.frame} with 199 rows and 26 columns.
}
\usage{
available_genomes
}
\description{
A data frame of genome info from UCSC
}
\keyword{data}
