% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_bam_assembly}
\alias{check_bam_assembly}
\title{Check bam assembly}
\usage{
check_bam_assembly(bam_file, assembly)
}
\description{
Checks bam file header against chrom.sizes for 'assembly' from UCSC. Returns logical.
}
