% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chrom_sizes_list}
\alias{chrom_sizes_list}
\title{A list of chrom.sizes from UCSC}
\format{
An object of class \code{list} of length 206.
}
\usage{
chrom_sizes_list
}
\description{
A list of chrom.sizes from UCSC
}
\keyword{data}
