% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_available_genomes}
\alias{get_available_genomes}
\title{Get available genomes
Helper function to return data frame with info on available genomes from UCSC.}
\usage{
get_available_genomes()
}
\description{
Get available genomes
Helper function to return data frame with info on available genomes from UCSC.
}
