% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_chrom_sizes}
\alias{get_chrom_sizes}
\title{Get chrome sizes
Helper function to return list with all available chrom.sizes.}
\usage{
get_chrom_sizes()
}
\description{
Get chrome sizes
Helper function to return list with all available chrom.sizes.
}
