% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSeq.R
\name{RSeq}
\alias{RSeq}
\title{RSeq Pipeline}
\usage{
RSeq(
  mode = NULL,
  experiment = NULL,
  control = NULL,
  operations = "all",
  outdir = "RSeq_out",
  analysis_params = NULL,
  samples = NULL
)
}
\arguments{
\item{mode}{Required. The \code{mode} parameter controls the workflow used to analyze data based
on best-practices for the specified R-loop mapping method. This can be overwritten
by manually passing parameters to \code{analysis_params}. Options include
\code{DRIP}, \code{DRIPc}, \code{qDRIP}, \code{ssDRIP}, \code{RDIP}, \code{R-ChIP},
\code{RR-ChIP}, and \code{MapR}. These options can also be specifid in \code{samples}.
Please read the vignette for more details.}

\item{experiment}{Character cotaining the GEO or SRA accession(s) (GSM or SRX) of experimental samples
to download and analyze. Can also be the file prefix of local fastq or bam file(s). Must be specified
in \code{samples} if not provided here.}

\item{control}{Same as \code{experiment} but for control/input samples.
Provide in the same order as \code{experiment} samples so that there is a 1:1 match
between each \code{experiment} sample and the corresponding control. Can be specified in \code{samples}.}

\item{operations}{Character specifying which analysis steps should be included. Default is \code{'all'} which
includes \code{c('trim&filter', 'align', 'coverage', 'peaks', 'QC', 'annotate', 'report')}. Can be specified
in \code{samples}. See \code{details} below.}

\item{outdir}{Character specifying folder locations to save data to. Default is \code{'RSeq_out'}. Can
be specified in \code{samples}.}

\item{analysis_params}{Named list specifying additional parameters to pass to each analysis step.
Can be specified in \code{samples} instead. See \code{details} below.}

\item{samples}{Data frame containing sample info and analysis parameters. See \code{details} below.}
}
\value{
An RSeq object containing \code{params}, \code{QC}, \code{report}, and \code{peaks}. See \code{details} below.
}
\description{
A flexible pipeline for R-loop mapping
}
\details{
\subsection{The RSeq Workflow}{
  To offer flexibility, the \code{RSeq} pipeline allows users to fully control the workflow parameters.
  This can be done simply by specifying \code{operations} and \code{analysis_params}

  \strong{operations}:

  The default pipeline for each \code{mode} follows the same pattern:
  \describe{
    \item{trim&filter}{Fastq reads are trimmed and filtered. Performed using \code{fastp} by default, which also provides an HTML report of read quality.}
    \item{align}{Alignment of reads to genome. Uses \code{bwa mem} by default with default params.}
    \item{coverage}{}

  }


}




\code{RSeq} can be used with or without specifying a \code{samples} file. Users will file that this file
provides convenience for complicated analysis. The 'samples' file allows complete flexibility with respect
to analysis parameters.

\section{Simple Analysis: without 'samples' file}{
  e
}
}
\examples{
genesOfInterest <- c("ATM", "SLC7A11")
correlationAnalyzeR::analyzeSingleGenes(genesOfInterest = genesOfInterest,
                              returnDataOnly = TRUE,
                              GSEA_Type = "simple",
                              Sample_Type = c("normal", "cancer"),
                              Tissue = c("respiratory", "pancreas"))

genesOfInterest <- c("BRCA1")
correlationAnalyzeR::analyzeSingleGenes(genesOfInterest = genesOfInterest,
                              GSEA_Type = "simple", returnDataOnly = TRUE,
                              crossCompareMode = TRUE,
                              whichCompareGroups = "normal")

}
